scripts/requirements.txt
psycopg2-binary
python-dotenv
finnhub-python
requests

